testing the animations
although the animations "work," since it technically uses it as a
seperate object and not a sprite 2d object there is no sprite to make it
flip if I want to walk left. So, the image disapears when walking the
other direction....

easy fix is for the artists to export to animation sequences to png
format since it's easier and we have to code the animations any way.
Additionally, doing it that method we don't have to amend our code or
problem solve to fix it. They can conform to what we have.